Load text from the FileDocumentManager
The document manager will strip CR from the file content. Since the
PSI is based on the text from the document manager we would get
test failures on Windows with the old code which would keep the CR.

The offset found via PSI would then not match with the text.

Bug: N/A
Test: LayoutInspectorTreePanelTest and others.
Change-Id: I5c6bdf5d589443d597659f3a91fd4fd0b0b04655

GitOrigin-RevId: 1234fee81a73feaaebd18dd0da45ef646e2a4c8e