[lint] Rewrite getReferenceExpressionNames
The previous implementation was specific to the number 3, which was
reasonable for most downstream uses of well-formed Gradle Dsl, but is
not technically correct even for simple cases as we could implicitly
be operating on arbitrary objects which happen to have properties or
methods of the same name.

The new implementation recurses all the way through the expression,
though does not change the behaviour on anything that it does not
understand in this CL; a subsequent CL will distinguish between early
and normal termination.

Bug: 279885639
Test: existing
Change-Id: I8d6f99ee4fdae9fe1f740e879263c7aadc19c5b7

GitOrigin-RevId: d1c9c6fcb2806587c6fd7146e672890a7e6329dd