Use module context to determine Gradle tasks
When invoking Gradle from a holder module, we should execute
its tasks in addition to its submodules'. When invoking Gradle
from sub-modules (ex: Main, UnitTest, AndroidTest) then we
should only invoke tasks on the sub-module itself.
This change uses the module type to determine the tasks to
execute. This prepares for the deprecation of TestcompileType.

Bug: 310177152
Bug: 235567998
Test: added
Change-Id: Id2e2deda0bc2e73fb8ae44ac8cd713a73bfb2265

GitOrigin-RevId: f81fd9fc2f8e1c4ac136e75685aa873c3937fb75