Make 'testOldKotlinDetectedAppliedInPluginDsl' able to run
This test was ignored because to make it run in gradle kgp plugin marker
is missing in checked-in artifacts. It does not make much sense to add
it just for this test though. To cover this particular test case we can
instead pretend that current kgp version (which does have plugin marker)
does not support Configuration Cache.

In order to make plugin resolution work there should be pluginManagement
section in test project settings.gradle. To not mess up with default
SimpleProject that is used everywhere create an adjusted copy of it just
for this test to build-attribution/testData folder.

Fixes: 303118113
Test: adding test
Change-Id: I14e84aa64608dacc8955a4fe9b9e74995ff137af

GitOrigin-RevId: 25a31a91842b0a172795546963f5a48a1a17e1a9