[android] IDEA-314901 Move AndroidComponentDownloader to intellij.android.common module
AndroidComponentDownloader downloads Android plugin resources in runtime, and it resolves path to these downloaded resources.
Modules that use resources downloaded in runtime are:
 - intellij.android.adb
 - intellij.android.profilers
 - intellij.android.profilersAndroid
 - intellij.android.transport
 - intellij.android.native-symbolizer

Since AndroidComponentDownloader was defined in intellij.android.core module, it couldn't be used in other modules that require it.
Because of this AndroidComponentDownloader is moved from intellij.android.core => intellij.android.common module.

GitOrigin-RevId: 3e65c3fb4b88a07a518317da9b8d5902939a2127