[PSD] Guava version number handling in Suggestions tab
In order to correctly determine whether there is an available update
for guava artifacts which have a family encoded in their version
number, we need to communicate more than just the latest version (and
latest stable version) as computed by version comparisons.  Therefore,
persist the complete list of versions available in an
AvailableLibraryUpdate object.

When making a recommendation in the project structure dialog for an
artifact in group com.google.guava whose version currently ends in
-android or -jre, filter the available versions by suffix and offer
the latest found.

Fixes: 315310265
Test: manual
Change-Id: I56e7277f2a696d566517894dbb10815299481c9c

GitOrigin-RevId: 52ff7c7479de9ef812cd774e673aa5b809abb78a