Refactor updateRoots -> refreshChildren
The implementations of `MultiResourceRepository` already have logic to
update themselves after project roots change, in their various
`updateRoots` methods. Most of these methods require
`StudioResourceRepositoryManager` to calculate something first and pass
it to `updateRoots`.

This change refactors the logic so that the update methods do not need
any parameters, instead calling into `StudioResourceRepositoryManager`
to fetch the same information. This will allow in subsequent changes
calls to update the repositories' children in a context that isn't
coming directly from the manager.

As part of this, I'm updating the naming to `refreshChildren`. The
previous `updateRoots` was in response to the trigger (project roots
were updated); the new name represents what the method does, regardless
of why it's called.

Bug: 313309846, 304316363
Test: Covered by existing
Change-Id: I22fba45a7fd3f278ca7fc23dcb249b3962c56946

GitOrigin-RevId: 7273ba5faddec74d93671d55da4e5aa8b48728ff