Fix cannot import module dependency for KMP projects
Due to KTIJ-28110 the kotlin IDE plugin is not able to map the
dependency sourceSet to the gradle sourceSet data node after we
resolve dependencies from a kmp module on an android module to
the appropriate sourceSet.

A new sourceSetDataNodeMap was introduced for the workaround
as an optimization. However, since we are reusing the same
instance of GradleSourceSetData for different opened KMP projects
will result during ModuleDependencyDataService.importData
to not be able to find the linked IDE module for the source set, since
the IdeModelsProviderImpl.myIdeModulesCache map was populated with
unmatched instance by the platform for the same source set, throwing:

"Can't import module dependency for '%s' module. Reason:
target module (%s) is not found at the ide"

Bug: 310044427
Test: AndroidTestConfigurationProducerTest with KGP 1.9.21
Change-Id: Ie41809faa1dbfac6d19ea9363b1662f30b742d21

GitOrigin-RevId: 3c10a50b8279cd9f3fabc6ef79c96ed52d27b945