Refresh child repositories when one is disposed
When a module's repositories (app, project, or module) are disposed,
they may still be referenced by another module. This doesn't happen most
of the time, just by luck. (Most of the time when one of these is
disposed, we are disposing all the repositories for all modules in
sequence. A stale reference only occurs when a resource repository is
accessed by another thread between two modules being refreshed; so it's
common enough to show up, but doesn't happen most of the time.)

This change ensures that when a repository is disposed, it notifies any
parents of that happening. The parents then are able to refresh their
own children, which gets rid of their reference to the disposed object
(and replaces it if necessary).

Bug: 313309846, 304316363
Test: StudioResourceRepositoryManagerTest.testDependentModuleReset
Change-Id: I00f3453fa04f30885c0dba9b273961eeb096418b

GitOrigin-RevId: a194e34e46a6c7e716d5fa459dd530ac31ae71af