Clean up kotlin-stdlib descriptor hack.
This workaround was required to inject the kotlin stdlib into any K2
test that performed Kotlin analysis, because analysis would crash if
classes from the stdlib were not available. That crash has been fixed in
kotlinc 1.9.20, which merged in with IJ 2023.3, so these hacky
workarounds are no longer required.

Some tests still require the Kotlin stdlib because the analysis they
perform relies on the stdlib classes being present, so ensure that those
tests actively opt in to using the Kotlin stdlib descriptor.

Bug: 294248298
Fixes: 300170256
Test: Treehugger + manual studio-linux-k2 regression test run
Change-Id: Ie2d9d84a0de96e932056b9ba2ccee878184a50ca

GitOrigin-RevId: 4f93db43134de59955a7210f1ad96539cdfeb6e2