Revert "Disable dm installation (Iguana Beta only)"
This reverts commit db0942ef135a503e3c672eec02213b1ba0c9ec9b.

Reason for revert: We unfortunately only had post-submit running on (iguana-dev) on Dec 13 and not presubmit. Looking the below Test Hub information after this CL was merged we started seeing significant test failures beginning with ab/11215808. Presubmits are now ON on the branch and we can resubmit this CL after presubmits pass.

https://android-build.corp.google.com/test_hub/search/?config=devtools%2Fstudio&filter=mediumFailureRate&filter=highFailureRate&invfilter=unhealthyInvocations&invfilter=staleInvocations&query=target%3A%28studio-linux%7Cstudio-win%7Cstudio-coverage%29+config%3Aandroid_studio%2Fbazel+branch%3Agit_studio-iguana-dev&tab=tests&useNewTable=true&newTableOrder=0%2C1%2C2

Change-Id: I30d768432426134b49c8d697542b330629a898eb

GitOrigin-RevId: f2b3f66fcd154fc779ae0a227bb4e2e8248faef2