Prevent auto session selection in tbux's updateSessionItemsByGroup
...so that soft restarting the Android Studio instance (just closing
the window and not the entire process) does not try to re-select
a default task automatically. However, there are two cases in which
auto-selection is enabled in the Task-Based UX:

(1) A new session was found, and it is ongoing.
(2) An existing session was found, and it has completed.
Both cases (1) and (2) require that there be a non-UNSPECIFIED
`currentTaskType` set and that the session was not imported.

Fixes: 315037689
Test: added
Change-Id: I8ee135727b5b9f7ca6a2ae88deb20a25365f8d8a

GitOrigin-RevId: 88bf33d943a44bd5d40f08ea7b92f2a5a000ec6f