Move loading of ResourceResolver to background
This change follows the same logic as with lazy loading of the color
resource values. If the ResourceResolver is not immediately available,
the value is returned without being resolved and a background update is
triggered.
Once the ResourceResolver is available, a new update of the model is
triggered to refresh the values depending on it.

Fixes: 316554064
Test: LazyCachedValueTest, NlPropertyItemTest
Change-Id: I383b6e238d305ae76b1ec08092f83128a8cab5ee

GitOrigin-RevId: 3e5be03d0d8f96f2325ad10d6e992a53751df4ae