Don't reset parents on facet disposal
ag/25629717 recently added logic to allow parent repositories to be
refreshed when one of their children is disposed. This seems to be
working as is; but a subsequent change is causing a deadlock on
disposal.

This change is reducing the scope of when parent repositories are
explicitly refreshed. There are two scenarios when child repositories
are disposed:

1. Child repos are disposed when resources are being explicitly reset.
   This is the scenario where we need to refresh parent repos, and the
   adjusted logic will still do this.

2. Child repos are disposed when the containing facet is disposed.
   Parent repos do *not* need to be refreshed in this case, as a
   different event will be raised that causes them to refresh. (Or the
   entire project is shutting down, in which case they still don't need
   to be refreshed.)

Bug: 313309846, 304316363
Test: StudioResourceRepositoryManagerTest.testDependentModuleReset
Change-Id: Ie14b58d46749c5e18dba1e2a69b58c37f33ae673

GitOrigin-RevId: f9557adb0b49bb6845ac1be1825716c3ea524035