Disable two asserts in AddComposableAnnotationQuickFixTest
...since Compose compiler 1.5.6 reports COMPOSABLE_EXPECTED error
on NonComposableFunction while it should only do this for a lambda
taking @Composable as an argument.

This scenario is not common because the only way to use a composable
inside of non-composable is to assign it to `mutableStateOf` which
is not common. And in this case a user will be presented with two
options to apply quickfix and they would have to choose the right one.

Bug: 309364913
Test: AddComposableAnnotationQuickFixTest
Change-Id: Ieccb8d575f979105e22bf92ac7adc7d6731a4514

GitOrigin-RevId: 4a8fc36d1b85d25205bfb14cd3624f8de7504a08