Always use a PropertyLabel for a text editor in Inspector
The Layout Inspector is never editing text in the property panel.
When a text editor is used in a variable height cell editor the
property panel may flicker because of the height difference
between a renderer (PropertyLabel) and an editor (PropertyTextField).
Remove this flickering by always using a PropertyLabel in the
Layout Inspector.

Fixes: 318538413
Test: ResolutionStackEditorProviderTest
Change-Id: I1b1e8da54d0d737226c2a060f4d531d288d998f6

GitOrigin-RevId: de1912c066b9f2ea1c861ff110c1756d250734c8