Adjust exportable criteria to include checking for child artifact
The criteria for exporting an imported trace was that the session
type was either MEMORY_CAPTURE or CPU_CAPTURE. This was not
sufficient as one assertion made in the export functionality is
that a single child artifact exists. Thus, the check for the single
child artifact has been added to the canExport criteria. Also, the
field of `canExport` has been converted to a getter so that if the
child artifacts are updated after initialization of the class's
instance, it will reflect that.

Fixes: 318479145
Test: added
Change-Id: I09e568e3c12a1fddd374ef3d3ee31e100a903519

GitOrigin-RevId: b891b2175e61409205a65d28a759d63cb513c290