Define a trampoline for Studio in the IDE
IntelliJ IDEA's Application Run Configuration, when it is requested to
use a module's classpath for the application, must choose which
classpath: one which includes only production dependencies, or one
which also includes test dependencies.

IDEA's logic for how to choose this is somewhat lossy when the main
class is not directly present in the module, but is instead in a
library; if there is an explicit test-scoped dependency to that
library, then the answer is obvious, but in other cases it falls back
to a heuristic involving whether the module in question has production
sources at all.

The `studio` module, before this change, had no production sources.
Consequently, the classpath generated for the Android Studio (and K2)
run configurations included test dependencies as well.  Defining this
trampoline convinces IDEA not to include test dependencies, which
should allow faster build-and-run times for developers working
primarily in the IDE.

Bug: 317101189
Test: manual
Change-Id: Ie88d5e78c6663abcdc17300c3b26f2fa5df81e9c

GitOrigin-RevId: 6285d022d61802ba481c14d1faddc388eb697f92