Add and use new Studio Bot APIs in Studio
This CL adopts a suite of APIs in the ml-api module that let clients
interact with Studio Bot in safe and expressive ways. The main entry
point is StudioBot, from which instances of the chat and llm services
can be obtained. All queries must adhere to the user's settings and
aiexclude files where applicable, and this is enforced by the APIs.

Bug: 315384497
Bug: 314179037
Bug: 310696988
Test: existing/manual
Change-Id: Ic0be06cb4b63d9640359c4a6b42d3aa14a59b2ff

GitOrigin-RevId: 8c3c70811575330ef71463a6ee38f7ce01953ec0