PopulateCachesActivity: start with isDumb()=true
Instead of queueing a DumbModeTask for later execution, we can run
PopulateCachesActivity in dumb-mode directly.

This change addresses a warning (in idea.log) from IntelliJ about
migrating this StartupActivity to ProjectActivity. As a bonus, this
also ensures that a smart-mode phase cannot sneak in between the
initial dumb-mode and PopulateCachesActivity.

Bug: 310034148
Test: existing + verified this code still runs at expected time
Change-Id: I753d3982fb7f44c6dcd1a189dd83e18a6fea59ed

GitOrigin-RevId: c4d1e8fdc67997d58f4bb502d28bb134278a0091