Fix flaky test in MainMemoryProfilerStageViewTest
MainMemoryProfilerStageView class object needs to be declared in
testWhenSessionDiesRecordingOptionsViewIsDisabled testcase for it
to hold a reference. If not, when object is being used in
addDependency method in AspectModel, it will be removed by
WeakHashMap since the object doesnt hold any reference.

This fix was tested in linux remote with 5000 runs and all passed.

Bug: 301024803
Test: Existing
Change-Id: I791356b2440b39e9ca9be2b4aa0f88a12e7b7b8d

GitOrigin-RevId: f2162797589fa64664b21b8d925b2131dfc8b79f