Handle app connecting to ADB before LE starts
Because Live Edit doesn't set up listening for ADB events until
notifyAppDeploy() is called with an app ID to listen to, it is possible
that the app may connect to ADB before this listener is set up.

To handle this case, we simply check if the app is already connected
before setting up our ADB listener, and set the LE status appropriately.

Fixes: 318739081
Test: Manually introducing delay before notifyAppDeploy() is called
Change-Id: Ib7a3ebc0d8b9927befab7e5e304b94ec0be337d0

GitOrigin-RevId: db4be90bc98ac21abdadc1316307f37e4768f67c