Revert to platform's setup of project libraries
Custom implementation had a bug where caching is
effectively disabled as Key was created on every
invocation. While it is possible to improve platform
implementation of this code, the safest thing to do
is to entirely revert to it. This might cause more
libraries to be "module-level" (instead of "project-level")
but because roots are shared between them, the impact
should be small.

Bug: 319624143
Test: manual performance tests
Change-Id: I0586dce021548b6f423b4b3c8860cb13e1c56222

GitOrigin-RevId: 97b44edffca2fc78ed66c73458209def9803e99c