Refactor IssuePanelService to make better use of IJ APIs
The core of this change is making DesignerCommonIssuePanel a
ProblemsViewTab. This allows to use the Design Tools custom tabs as any
other Problems window tabs by leveraging IJ APIs.
In addition, DesignerCommonIssuePanel is also made into a
SimpleToolWindowPanel which enables it to react to orientation changes
of the Problems window (e.g. going from bottom to left of IDE) in the
same way as the other Problems tabs.
This change makes it possible to handle the creation of the tabs through
ProblemsViewPanelProvider, one for the shared panel and one for UI check
panels. The one for the shared panel is registered as a plugin
extension, which means that the panel creation happens automatically on
loading the plugin instead of having to be created by the
IssuePanelService.
The rest of the modifications are direct consequences of the changes
mentioned above.

Bug: 319640923
Test: tests updated
Change-Id: Ia457ad94505e4932416f156dc94028a300af79a0

GitOrigin-RevId: f911d2d17fa19b7b4e4b3bb4de1da3089d2c00b6