Migrate Gradle StartupActivities to ProjectActivities
We need to respect the ordering between the activities, formerly
semi-implicit in the plugin xml files: we must wait for any New
Project Initialization to complete before attempting to attach models
or trigger Sync (and indeed before querying whether the project
isBuildWithGradle), and we must wait for models to be attached, if
they will be attached, before setting up the Native workspace.

Provide an internal project service for each activity with dependents,
which allows those dependents to wait for the completion of the other.
In AndroidGradleProjectStartupActivity, additionally wait for
JpsProjectLoadingManager and ExternalProjectsManagerImpl to finish
initializing.

The existing test of AndroidGradleProjectStartupActivity needs to be
rewritten to adjust to the new threading discipline.  In particular as
a unit-like test it should not run on the EDT without a coroutine
scope, because that interferes with the coroutines' ability to
interrupt the EDT.

Bug: 287942576
Bug: 310034148
Bug: 319261172
Bug: 319461907
Test: manual, existing, adjusted
Change-Id: Ic89c558e48b51672700682b53452580d9590e8e9

GitOrigin-RevId: 455bac4617973b40ad1947a970a43e770ff37827