Update tooltip for Layout Validation on config change
The label above each preview in layout validation has a tooltip
summarizing the configuration of the preview. When a configuration
change happens, this needs to be updated.
This also removes the tooltip as a constructor argument for NlModel
as it is almost always set to null. Instead, this creates a setter
method for it.

Fixes: 319214857
Test: test added
Change-Id: I8351b435af1c1e1ecbcdd011cfb685de898bd174

GitOrigin-RevId: cca9dc7599812197e0ce25cf5759365cc1bbe6f5