Replace hasAndroidFacets with isAndroidProject
as it is the intention of all usages.

Make `hasAndroidFacets` deprecated and log an error to prevent
accidental usages in Android Studio.

`isAndroidProject` is likely used by IntelliJ itself so keep this
abstraction as is but keep it possible to change the implementation in
Android Studio only. Other Android Studio project systems may need a
different implementation.

Bug: n/a
Test: n/a
Change-Id: Iad9ec36c194e47bd168ebde54441b7495f0c582c

GitOrigin-RevId: 2102ed2b6af17861e7ce5e91335fa46ed2a993ab