Invert transforming variable versions in testing for golden files
Instead of transforming the particular version that we have used into
a marker string, then comparing with the golden file, transform the
golden file's context to insert the particular version that we used
instead of the marker string.  This means that we will not get
spurious failures in the event of a Kotlin version (or androidx
core-ktx version) happening to collide with some other version number
or even other string in the system.

Bug: n/a
Test: this is tests
Change-Id: I83fd3696e065f815e7e73464caf19e920647cf90

GitOrigin-RevId: d01bfa52baa1ff16099a5ec7b015407a0a9574bf