Snapshot changes for dependency graph traversal relaxation
In the accompanying AGP change, we stop querying artifacts
for Android project dependencies in FullDependencyGraphBuilder.
Formerly, except for some special cases, we would traverse
through nodes in the graph only if they have an artifact.
But with these changes, no Android project dependency will have
a corresponding artifact, so we have to relax this.

That means, some behaviour related to dynamic features change,as
we rely on the fact that there is no AAR published from an
app to detect an app dependency and cut the app's subgraph from
dynamic feature to reflect it more closely to the packaged reality.
But without an artifact it's not possible to do this anymore.

However, that's fine because the IDE shouldn't actually care about
what's packaged, instead it should care what the dynamic feature
will experience. In all the cases we use the dynamic feature's
classpath, we should actually merge it with the base app's classpath
first, anyway.

I say "should", because this is not actually currently the case
for AndroidGradleClassJarProvider, and this actually breaks
rendering when a feature uses custom view through the app.
I confirmed the corresponding AGP change actually fixes this case.

AndroidGradleClassJarProvider should ideally always include app
classpath for dynamic features, I plan to send that in a follow-up
change.

Note: This only modifies a privacy sandbox snapshot test, although
that's just a coincidence, it looks like this is the only valid
dynamic feature snapshot test present. We should probably look into
adding one.

Bug: 320453969
Test: privacy sandbox sdk snapshot test
Change-Id: I1845dc95f11516e5177516bf125192cfbdf7a966

GitOrigin-RevId: c896144679458ef5b5c57cf3a94ba776ec2315cb