Fix GuiTestLauncher on Windows
I suspect the introduction of PathClassLoader as the system class loader broke GuiTestLauncher on Windows. Bazel provides a classpath jar and GuiTestLauncher creates another classpath jar that just wraps the first. PathClassLoader assumes the name of a classpath jar starts with "classpath", which is a stronger assumption than URLClassLoader makes, and Bazel definitely does not follow this naming scheme, so the second
classpath jar is not correctly recognized.

This CL removes the indirection of classpath jars and normalizes the paths provided in the manifest. The normalization is again something not strictly required in Java, but oddly enough Gradle sync fails with mysterious errors about not being able to retrieve file attributes for those jars.

Bug: 274514598
Test: UI tests
Change-Id: Ia5ae185e0bf2ddf8720706257b0eac81024b1447

GitOrigin-RevId: 6690cf39a9b8bbe1d45dc3ab249ece26937720d0