Generalize getOrCreateDefaultResDirectory with a Token
Quite apart from the fact that this is unimplementable in some project
systems and has an unhelpful interface, the current code has at least
two bugs: if for whatever reason the project has no resource directory
findable by the source providers, we will return a File corresponding
to a possibly non-existent location on disk, from the default (Gradle)
module template; if the project has resource directories according to
the source provider but they do not exist on disk, we will return a
File corresponding to a file (not a directory).  Fix these issues, and
implement the Token for the Gradle Project System.

Bug: 147586218
Bug: 300434824
Test: existing tests adjusted, integration test added
Change-Id: Ifb88644bb4f9d9b7cd10e63a049a94e3282673e6

GitOrigin-RevId: d6a4144c99e379fd73d20d477b2d7272866b5a99