Remove more usage of MapDataContext
While arguably best done in another change, this also removes our
custom implementation of MapDataContext. SimpleDataContext does
not have the drawback of the platform's MapDataContext in that it
does not rely on the test framework, but only the action system.
I was just too lazy to start a new commit when I discovered this.

Bug: N/A
Test: Existing
Change-Id: I6b8b9f82a8e9e296992d32c91ee09314f2c5e63d

GitOrigin-RevId: 34e9234a3762588c27c963eb1264c1b1663c5ba1