Do not leak DesignerCommonIssuePanel in Disposer
When adding listeners, we register their removal in the Disposer, in
case the object they are linked to gets disposed before the panel. That
avoids leaking disposed objects through the listeners. But we did not
dispose of those removal disposables when disposing the
DesignerCommonIssuePanel. This means that the disposer can hold on to a
disposed panel unnecessarily. This change makes sure that both cases
are handled correctly.
This also sets the correct parent dispoable to the
DesignToolsIssueProvider, to make sure that it does not stay behind when
cleaning up the DesignerCommonIssuePanel.

Bug: N/A
Test: N/A
Change-Id: Ifeae8c901b40dd52a7ead007f30299f9fdc11375

GitOrigin-RevId: 8cb546ec83c0cff0455c55ff203853c445541af3