Workaround for runReadActionWhenSmart UI deadlock
Invoking runReadActionWhenSmart with the read lock can cause a deadlock
and the method guards against that misuse. However, after IntelliJ
2023.3, calling without the read lock and in the UI thread can also
cause one.
The method should never be invoked in the UI thread but we have a number
of calls that do that. This adds a new utility method that guards
against that situation and avoids calling the method in the UI thread
without the read lock.

Fixes: 322507246
Test: ThemeUtilsIndexQueryTest
Change-Id: Ifbae9a2bead599f5b3a5bb9bb13b11b5318b30d1

GitOrigin-RevId: d8012ecef1d0a6d5c3510cc8b7195b98c930aa8b