Invalidate AAR caches *before* a library is unloaded
K2 has stricter ordering requirements on cache invalidation due to
somewhat more precise cache invalidation strategies: when a module is
disposed of, the cache invalidation event should be emitted before that
to not have invalid modules in FIR session caches for even a split
second. Doing it in the wrong order can trip up some assertions.

Bug: 322954494
Test: BuildVariantsIntegrationTest
Change-Id: Id7022535867186443cb56dfcd793473a805ead11

GitOrigin-RevId: 5c5f2ebfb902b362fa3ee506706ac8280672710e