Do Not Show Unsupported Content As Text
This broke when `IntellijDataViewer` was converted to Kotlin.

It seems that the reason the old code worked is because of an exception
that was thrown while trying to create a PRETTY viewer for unsupported
content.

The fix handles this better.

Note that `ContentType` enum contains 3 distinct groups:
1. Image types: BMP, GIF, JPEG, JPG, PNG, WEBP
2. Text types: CSV, HTML, JSON, XML
3. Other: DEFAULT

Also note that `text` is also supported even though it's not included
in the enum.

Fixes: 322750058
Test: Added
Change-Id: Iacbd85f2b140585c410ba67ea6091766c26977c6

GitOrigin-RevId: 3e149f957ba9421f43bc8924206767fbcf8812ef