Take PSI snapshots on run/file open
Previously, we took PSI snapshots for LE validation in the moment before
a document change event occurred; however, some editor changes (color
picker, for one) modify the PSI *before* any document events are fired.

In these cases, taking the snapshot before the change is too late, as
the PSI changes have already been committed, and the diff will
erroneously return "no changes". This causes PSI-first changes to not be
validated.

To fix this, we take snapshots 1) of all open files when the app is run
and 2) when a new file is opened/editor window is switched to.

Fixes: 318036046
Test: PsiValidatorTest, Existing
Change-Id: Ibaf5419f5224506cf2a327b8fe20b98fcaf003ba

GitOrigin-RevId: 60b06e95703637c857a6fecb7e3b1452d242a0b7