Wrap getModuleTextWithIcon in readAction
This currently called from the EDT in the platform. As part of 2023.3
changes, all readAction's (even if on EDT) should be wrapped in
readActions causing the user visible RuntimeException. Since there's
several calls in the platform it seems easier to wrap our call for now
instead of fixing all the calls in the platform.

Fixes: 323566810
Change-Id: I12454c00cf9f525537f723924ba1e93ae6cc6b04

GitOrigin-RevId: 7d957c7df30fd733e6a52b74f004d770aa7c8055