Cancel pending render requests when deactivating
...LayoutlibSceneManager.

This CL fixes a potential blockage that can occur when calling
SceneManager#render. If the request is enqueued right before the scene
manager is disposed/deactivated, the caller will stay stuck waiting for
the render to complete. The render will never complete as the scene
manager is disposed of/deactivated.

Instead, with this CL, we ensure any pending renders are completed
exceptionally to unblock any callers.

Bug: N/A
Test: LayoutlibSceneManager
Change-Id: I1ad583940bc1cc680c210bb7264125b17773d61d

GitOrigin-RevId: 2bf429a291a5501dbd2cb36dcd132c2baeaca937