Ensure a single scene manager is created when adding
... a duplicated model concurrently.

This CL fixes a potential race condition in which two scene managers can
be created for the same model. This leads to the extra scene manager
not being referenced by the DesignSurface anymore, which would
potentially lead to further issues. For example, this extra scene
manager will not be disabled as it should when removed in
DesignSurface#removeModelImpl.

Bug: N/A
Test: DesignSurfaceTest
Change-Id: If4cdaab7c05a1b629373deb6c2735a178ffe9634

GitOrigin-RevId: 3b8ad44a21308edb0f6b5ed35108e83ee5167170