Fix flaky test: AppInspectionPropertiesProviderTest
The test was relying on the number of callbacks from property
generations. Unfortunately the callback was added while some
background threads was working on connecting to a device which
would also generate 2 property events.

The flakiness has been removed by setting up the callbacks before
trying to connect to the device.

Fixes: 323251521
Test: AppInspectionPropertiesProviderTest
Change-Id: I66f3648c4841bb9c9ab417a72160c04253f726b8

GitOrigin-RevId: ba9eb35d301155ef697696d5e65870326d0c1b93