[lint] Be more precise about the root directory in the IDE
In the IDE, lint often operates on subsets of the entire project.
This means that the default implementation of LintClient.getRootDir(),
which effectively takes the largest common prefix of the modules the
particular invokation of lint is analysing, will fail to find the
correct root if all modules are (strictly) nested under a single outer
module.

However, in the IDE we can query the system directly for the known
root of any given module, so do that instead, and fall back to the
default implementation if that fails.

Fixes: 324183393
Test: added integration test
Change-Id: I87cea9833382b387b5bdf9746b0b363ef8188ee3

GitOrigin-RevId: ba9e1706020b73358dbbd266f6e729b07c7aecb6