Don't cause stub creation in letIfLiveEditable
Checking for isScript() causes the stub to be loaded/created if it
hasn't been already. Doing this here is wrong for several reasons, but
most pressing of which is it causes tests surprise load stubs when they
weren't expecting it.

In reality, *all* of the heavy operations in LiveEditServiceImpl should
be moved elsewhere, but this is a start.

Bug: None
Test: Verified that test suites do not end with stale file ids
Change-Id: Ia8c0135ab6581206b86af01f4d1b6d27f30ef739

GitOrigin-RevId: df4f2a271c09cc1a4402eda098b009ee17bb6d8e